update to new opal_infosubscriber_t structure
There was a restructuring of the comm/win/group objects
that shifted the fields down for getting at the class name.
This appears to relate to opal_infosubscriber_t changes from 50aa143.

NOTE: The opal_infosubscriber_t changes were not applied to the
oshmem_group_t structure, so no need for changes there.