WIP: add test program for opal memprof bits
NOTE: I decided that it was a bad idea to add a 'simple_memprof' driver
under the test/opal/ subdir, b/c it will need MPI, which is not required
for OPAL.  And does not seem right to add 'memprof' subdir with just one
test.  So for now, just leaving it out.

TODO: Need to clean all this up in my branches and back out the
prelim test code added to meminst-master and meminst-v4-0-x.